requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
asyncssh==2.17.0
atpublic==5.0
attrs==24.2.0
bcrypt==4.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.1
dvc==3.55.2
dvc-data==3.16.6
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.1
einops==0.8.0
entrypoints==0.4
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
graphviz==0.20.3
grpcio==1.67.0
gto==1.7.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.28.0
iterative-telemetry==0.0.9
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kombu==5.4.2
lightning-utilities==0.11.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.11.0
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pygit2==1.15.1
Pygments==2.18.0
pygtrie==2.5.0
PyNaCl==1.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.2
scipy==1.14.1
scmrepo==3.3.8
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
simsimd==6.0.3
six==1.16.0
smmap==5.0.1
sqltrie==0.11.1
stack-data==0.6.3
stringzilla==3.10.5
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomlkit==0.13.2
torch==2.4.1
torchinfo==1.8.0
torchmetrics==1.5.0
torchview==0.2.6
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.13.1
zc.lockfile==3.0.post1
zipp==3.20.2
